fix linter warnings
