+++ client/src/hooks/index.ts
@@ -1,7 +1,10 @@
 export * from './useAuth.tsx';
+export * from './useAuthProvider.tsx';
 export * from './useCohortsData.tsx';
 export * from './useDashboardData.tsx';
 export * from './useDebounce.tsx';
 export * from './useLocalStorage.tsx';
 export * from './useTraineeInfoData.tsx';
 export * from './useTraineeSearchData.tsx';
+export * from './useTraineeProfileContext.tsx';
+export * from './useTraineeProfileProvider.tsx';
