+++ server/package.json
@@ -10,7 +10,7 @@
     "build": "tsc && npm run sentry:sourcemaps",
     "dev": "nodemon -w src src/index.ts",
     "start": "tsc && node dist/index.js",
-    "lint": "eslint .",
+    "lint": "tsc --noEmit && eslint .",
     "lint:fix": "eslint --fix .",
     "setup": "npm install --include=dev",
     "dbsetup": "ts-node scripts/setup/index.ts",
